Refactor main show loop code all the way into Show struct.
