+++ tunnelclient/src/timesync.rs
@@ -49,6 +49,11 @@ impl Client {
         Client {socket, poll_period: Duration::from_millis(500), n_meas: 10}
     }
 
+    /// Return an estimate for how long a synchronization will take.
+    pub fn synchronization_duration(&self) -> Duration {
+        self.poll_period * self.n_meas as u32
+    }
+
     /// Take a time delay measurement.
     fn measure(&mut self) -> Result<Measurement, Box<Error>> {
         let now = Instant::now();
